refactor(frontend): split create_source into smaller mods
