+++ src/frontend/src/handler/create_source/external_schema/debezium.rs
@@ -1,0 +1,24 @@
+// Copyright 2024 RisingWave Labs
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+use super::*;
+
+pub async fn extract_debezium_avro_table_pk_columns(
+    info: &StreamSourceInfo,
+    with_properties: &WithOptionsSecResolved,
+) -> Result<Vec<String>> {
+    let parser_config = SpecificParserConfig::new(info, with_properties)?;
+    let conf = DebeziumAvroParserConfig::new(parser_config.encoding_config).await?;
+    Ok(conf.extract_pks()?.drain(..).map(|c| c.name).collect())
+}
